Started work on a new method for generating the BeamTransfer functions.
With this new method, we skip generating the frequency ordered files,
and instead skip straight to generating the m-ordered ones. It
generates subsets of all frequencies and baselines, and distributes
them across all nodes, then transposes to group them by m. It then
writes them all out to the m-ordered files. This is significantly more
parallelisable.