feat(beamtransfer): add robust check of whether beam SVD is complete per m
In the main loop over m, we check whether a beam SVD file exists, whether we can open it, and whether its
complete attribute is True. If either of those tests fails, we (re)generate the file and computations,
but if all tests pass, we skip that m in the loop.