feat(beamtransfer): if beam SVD pinv fails, try pinv2
scipy.linalg.pinv2() is an alternative to pinv() that uses an SVD-based algorithm instead
of a least-squares solver to compute a pseudoinverse. This commit gets the code to try
pinv2() if pinv() fails. If pinv2() works, an attribute is added to the beam SVD file
indicating which frequencies pinv2() has been used for.

Closes #94