fix(PolarisedTelescopeExternalBeam): include correct complex check for HEALPixBeam
The beam dataset in HEALPixBeam is a structured np.dtype that can't pass the np.issubclass_
check. This patch extracts and checks the types of the dtype fields explicitly.

This issue was noticed in ch_pipeline and fixed by Alex Reda.

black is also applied to the entire telescope.external_beam module in this commit.